@@ -29,7 +29,9 @@ import java.util.List;
 
 import hudson.util.VersionNumber;
 
-public record Plugin(String name, VersionNumber version, String scm, ZonedDateTime releaseTimestamp, List<String> labels) {
+public record Plugin(String name, VersionNumber version, String scm,
+                     ZonedDateTime releaseTimestamp, List<String> labels,
+                     int popularity, String requiredCore, String defaultBranch) {
     public io.jenkins.pluginhealth.scoring.model.Plugin toPlugin() {
         return new io.jenkins.pluginhealth.scoring.model.Plugin(this.name(), this.version(), this.scm(), this.releaseTimestamp());
     }
