@@ -32,16 +32,24 @@ import java.util.stream.Stream;
 import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 @Component
 @Order(DependabotProbe.ORDER)
 public class DependabotProbe extends Probe {
+    private static final Logger LOGGER = LoggerFactory.getLogger(DependabotProbe.class);
     public static final int ORDER = LastCommitDateProbe.ORDER + 1;
+    public static final String KEY = "dependabot";
 
     @Override
     protected ProbeResult doApply(Plugin plugin, ProbeContext context) {
+        if (plugin.getDetails().get(SCMLinkValidationProbe.KEY) == null) {
+            LOGGER.error("Couldn't run {} on {} because previous SCMLinkValidationProbe has null value in database", key(), plugin.getName());
+            return ProbeResult.error(key(), "SCM link has not been validated yet");
+        }
         final Path scmRepository = context.getScmRepository();
         final Path githubConfig = scmRepository.resolve(".github");
 
@@ -61,11 +69,16 @@ public class DependabotProbe extends Probe {
 
     @Override
     public String key() {
-        return "dependabot";
+        return KEY;
     }
 
     @Override
     public String getDescription() {
         return "Checks if dependabot is configured on a plugin.";
+    }
+
+    @Override
+    protected boolean isSourceCodeRelated() {
+        return true;
     }
 }
