@@ -40,6 +40,7 @@ import org.springframework.stereotype.Component;
 @Order(value = KnownSecurityVulnerabilityProbe.ORDER)
 public class KnownSecurityVulnerabilityProbe extends Probe {
     public static final int ORDER = UpForAdoptionProbe.ORDER + 1;
+    public static final String KEY = "security";
 
     @Override
     protected ProbeResult doApply(Plugin plugin, ProbeContext context) {
@@ -64,7 +65,7 @@ public class KnownSecurityVulnerabilityProbe extends Probe {
 
     @Override
     public String key() {
-        return "security";
+        return KEY;
     }
 
     public int getOrder() {
