@@ -25,12 +25,14 @@
 package io.jenkins.pluginhealth.scoring.probes;
 
 import java.io.IOException;
-import java.util.List;
 
+import io.jenkins.pluginhealth.scoring.config.GithubConfiguration;
+import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
 import io.jenkins.pluginhealth.scoring.model.ResultStatus;
 import io.jenkins.pluginhealth.scoring.model.updatecenter.UpdateCenter;
 import io.jenkins.pluginhealth.scoring.service.PluginService;
+import io.jenkins.pluginhealth.scoring.service.ProbeService;
 import io.jenkins.pluginhealth.scoring.service.UpdateCenterService;
 
 import org.slf4j.Logger;
@@ -45,17 +47,19 @@ import org.springframework.stereotype.Component;
  * Each implementation can choose to placed before or after another implementation using {@link org.springframework.core.annotation.Order} flag.
  */
 @Component
-public class ProbeEngine {
+public final class ProbeEngine {
     private static final Logger LOGGER = LoggerFactory.getLogger(ProbeEngine.class);
 
-    private final List<Probe> probes;
+    private final ProbeService probeService;
     private final PluginService pluginService;
     private final UpdateCenterService updateCenterService;
+    private final GithubConfiguration githubConfiguration;
 
-    public ProbeEngine(List<Probe> probes, PluginService pluginService, UpdateCenterService updateCenterService) {
-        this.probes = List.copyOf(probes);
+    public ProbeEngine(ProbeService probeService, PluginService pluginService, UpdateCenterService updateCenterService, GithubConfiguration githubConfiguration) {
+        this.probeService = probeService;
         this.pluginService = pluginService;
         this.updateCenterService = updateCenterService;
+        this.githubConfiguration = githubConfiguration;
     }
 
     /**
@@ -65,39 +69,84 @@ public class ProbeEngine {
         LOGGER.info("Start running probes on all plugins");
         final UpdateCenter updateCenter = updateCenterService.fetchUpdateCenter();
         pluginService.streamAll().parallel()
-            .peek(plugin -> {
-                try {
-                    final ProbeContext probeContext = new ProbeContext(plugin.getName(), updateCenter);
-                    probes.forEach(probe -> {
-                        try {
-                            final ProbeResult previousResult = plugin.getDetails().get(probe.key());
-                            if (previousResult == null || !probe.requiresRelease() ||
-                                (probe.requiresRelease()
-                                    && previousResult.timestamp() != null
-                                    && previousResult.timestamp().isBefore(plugin.getReleaseTimestamp()))
-                            ) {
-                                if (LOGGER.isTraceEnabled()) {
-                                    LOGGER.trace("Running {} on {}", probe.key(), plugin.getName());
-                                }
-                                final ProbeResult result = probe.apply(plugin, probeContext);
-                                if (result.status() != ResultStatus.ERROR) {
-                                    plugin.addDetails(result);
-                                }
-                            } else {
-                                if (LOGGER.isDebugEnabled()) {
-                                    LOGGER.debug("{} requires a release of {} to be processed.", probe.key(), plugin.getName());
-                                }
-                            }
-                        } catch (Throwable t) {
-                            LOGGER.error("Couldn't run {} on {}", probe.key(), plugin.getName(), t);
-                        }
-                    });
-                    probeContext.cleanUp();
-                } catch (IOException ex) {
-                    // TODO
-                }
-            })
-            .forEach(pluginService::saveOrUpdate);
+            .forEach(plugin -> this.runOn(plugin, updateCenter));
         LOGGER.info("Probe engine has finished");
+    }
+
+    /**
+     * Runs all the probes on a specific plugin
+     *
+     * @param plugin the selected plugin to run all probes on
+     * @throws IOException thrown when the update-center cannot be retrieved
+     */
+    public void runOn(Plugin plugin) throws IOException {
+        final UpdateCenter updateCenter = updateCenterService.fetchUpdateCenter();
+        runOn(plugin, updateCenter);
+    }
+
+    private boolean shouldExecuteProbe(Probe probe, ProbeResult previousResult, Plugin plugin, ProbeContext ctx) {
+        if (previousResult == null) {
+            return true;
+        }
+        if (probe.requiresRelease() && (previousResult.timestamp() != null)
+            && previousResult.timestamp().isBefore(plugin.getReleaseTimestamp())) {
+            return true;
+        }
+        if (probe.isSourceCodeRelated() && ctx.getLastCommitDate().map(date -> previousResult.timestamp() != null
+            && previousResult.timestamp().isBefore(date)).orElse(false)) {
+            return true;
+        }
+        if (!probe.requiresRelease() && !probe.isSourceCodeRelated()) {
+            return true;
+        }
+        return false;
+    }
+
+    private void runOn(Plugin plugin, UpdateCenter updateCenter) {
+        final ProbeContext probeContext;
+        try {
+            probeContext = probeService.getProbeContext(plugin.getName(), updateCenter);
+        } catch (IOException ex) {
+            LOGGER.error("Cannot create temporary plugin for {}", plugin.getName(), ex);
+            return;
+        }
+        try {
+            probeContext.setGitHub(githubConfiguration.getGitHub());
+        } catch (IOException ex) {
+            LOGGER.error("Cannot create connection to GitHub", ex);
+            return;
+        }
+        probeService.getProbes().forEach(probe -> {
+            try {
+                final ProbeResult previousResult = plugin.getDetails().get(probe.key());
+                if (shouldExecuteProbe(probe, previousResult, plugin, probeContext)) {
+                    if (LOGGER.isTraceEnabled()) {
+                        LOGGER.trace("Running {} on {}", probe.key(), plugin.getName());
+                    }
+                    final ProbeResult result = probe.apply(plugin, probeContext);
+                    if (result.status() != ResultStatus.ERROR) {
+                        plugin.addDetails(result);
+                    }
+                } else {
+                    if (LOGGER.isDebugEnabled()) {
+                        LOGGER.debug("{} requires a release of {} to be processed.", probe.key(), plugin.getName());
+                    }
+                }
+            } catch (Throwable t) {
+                LOGGER.error("Couldn't run {} on {}", probe.key(), plugin.getName(), t);
+            }
+        });
+
+        try {
+            pluginService.saveOrUpdate(plugin);
+        } catch (Throwable e) {
+            LOGGER.error("Could not save result of probe engine for plugin {}", plugin.getName(), e);
+        }
+
+        try {
+            probeContext.cleanUp();
+        } catch (IOException ex) {
+            LOGGER.error("Failed to cleanup {} for {}", probeContext.getScmRepository(), plugin.getName(), ex);
+        }
     }
 }
