@@ -25,19 +25,13 @@
 package io.jenkins.pluginhealth.scoring.probes;
 
 import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.time.Duration;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import io.jenkins.pluginhealth.scoring.config.GithubConfiguration;
 import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
+import io.jenkins.pluginhealth.scoring.model.ResultStatus;
 
-import com.sun.istack.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.annotation.Order;
@@ -50,27 +44,23 @@ import org.springframework.stereotype.Component;
 @Order(value = SCMLinkValidationProbe.ORDER)
 public final class SCMLinkValidationProbe extends Probe {
     private static final Logger LOGGER = LoggerFactory.getLogger(SCMLinkValidationProbe.class);
-    public static final String GH_REGEXP = "https://(?<server>[^/]*)/(?<repo>jenkinsci/[^/]*)(?:/(?<folder>.*))?";
-    public static final Pattern GH_PATTERN = Pattern.compile(GH_REGEXP);
 
+    private static final String GH_REGEXP = "https://(?<server>[^/]*)/(?<repo>jenkinsci/[^/]*)(?:/(?<folder>.*))?";
+    public static final Pattern GH_PATTERN = Pattern.compile(GH_REGEXP);
     public static final int ORDER = DeprecatedPluginProbe.ORDER + 20;
     public static final String KEY = "scm";
 
-    private final HttpClient httpClient;
-    private final GithubConfiguration githubConfiguration;
-
-    public SCMLinkValidationProbe(HttpClient httpClient, GithubConfiguration githubConfiguration) {
-        this.httpClient = httpClient;
-        this.githubConfiguration = githubConfiguration;
-    }
-
     @Override
     public ProbeResult doApply(Plugin plugin, ProbeContext context) {
+        final ProbeResult deprecatedProbeResult = plugin.getDetails().get(DeprecatedPluginProbe.KEY);
+        if (deprecatedProbeResult != null && deprecatedProbeResult.status().equals(ResultStatus.FAILURE)) {
+            return ProbeResult.failure(key(), "Plugin is deprecated");
+        }
         if (plugin.getScm() == null || plugin.getScm().isBlank()) {
             LOGGER.warn("{} has no SCM link", plugin.getName());
             return ProbeResult.failure(key(), "The plugin SCM link is empty");
         }
-        return fromSCMLink(plugin.getScm());
+        return fromSCMLink(context, plugin.getScm());
     }
 
     @Override
@@ -95,7 +85,7 @@ public final class SCMLinkValidationProbe extends Probe {
         return true;
     }
 
-    private ProbeResult fromSCMLink(@NotNull String scm) {
+    private ProbeResult fromSCMLink(ProbeContext context, String scm) {
         Matcher matcher = GH_PATTERN.matcher(scm);
         if (!matcher.find()) {
             if (LOGGER.isDebugEnabled()) {
@@ -105,22 +95,10 @@ public final class SCMLinkValidationProbe extends Probe {
         }
 
         try {
-            final HttpRequest request = HttpRequest.newBuilder()
-                .uri(URI.create("https://api.github.com/repos/%s".formatted(matcher.group("repo"))))
-                .timeout(Duration.ofSeconds(2))
-                .header("Authorization", "token %s".formatted(githubConfiguration.getGitAccessToken()))
-                .header("User-Agent", "Plugin-Health-Scoring")
-                .GET()
-                .build();
-            final HttpResponse<String> response = this.httpClient.send(request, HttpResponse.BodyHandlers.ofString());
-
-            if (response.statusCode() == 200) {
-                return ProbeResult.success(key(), "The plugin SCM link is valid");
-            } else {
-                return ProbeResult.failure(key(), "The plugin SCM link is invalid");
-            }
-        } catch (IOException | InterruptedException ex) {
-            return ProbeResult.failure(key(), "The plugin SCM cannot be validated");
+            context.getGitHub().getRepository(matcher.group("repo"));
+            return ProbeResult.success(key(), "The plugin SCM link is valid");
+        } catch (IOException ex) {
+            return ProbeResult.failure(key(), "The plugin SCM link is invalid");
         }
     }
 }
