@@ -61,6 +61,11 @@ class KnownSecurityVulnerabilityProbeTest {
     }
 
     @Test
+    public void shouldHaveDescription() {
+        assertThat(spy(KnownSecurityVulnerabilityProbe.class).getDescription()).isNotBlank();
+    }
+
+    @Test
     public void shouldBeOKWithNoSecurityWarning() {
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
@@ -83,7 +88,7 @@ class KnownSecurityVulnerabilityProbeTest {
     public void shouldBeOKWithWarningOnDifferentPlugin() {
         final String pluginName = "foo-bar";
         final VersionNumber pluginVersion = new VersionNumber("1.0");
-        final var pluginInUC = new Plugin(pluginName, pluginVersion, "scm", ZonedDateTime.now().minusHours(1), List.of());
+        final var pluginInUC = new Plugin(pluginName, pluginVersion, "scm", ZonedDateTime.now().minusHours(1), List.of(), 0, "", "main");
         final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
         final ProbeContext ctx = mock(ProbeContext.class);
         final KnownSecurityVulnerabilityProbe probe = new KnownSecurityVulnerabilityProbe();
