@@ -26,6 +26,7 @@ package io.jenkins.pluginhealth.scoring.probes;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -38,14 +39,19 @@ import java.io.IOException;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Stream;
 
+import io.jenkins.pluginhealth.scoring.config.GithubConfiguration;
 import io.jenkins.pluginhealth.scoring.model.Plugin;
 import io.jenkins.pluginhealth.scoring.model.ProbeResult;
 import io.jenkins.pluginhealth.scoring.model.ResultStatus;
+import io.jenkins.pluginhealth.scoring.model.updatecenter.UpdateCenter;
 import io.jenkins.pluginhealth.scoring.service.PluginService;
+import io.jenkins.pluginhealth.scoring.service.ProbeService;
 import io.jenkins.pluginhealth.scoring.service.UpdateCenterService;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -53,22 +59,36 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class ProbeEngineTest {
+    @Mock private PluginService pluginService;
+    @Mock private ProbeService probeService;
+    @Mock private UpdateCenterService updateCenterService;
+    @Mock private GithubConfiguration githubConfiguration;
 
-    @Mock
-    private PluginService pluginService;
-    @Mock
-    private UpdateCenterService updateCenterService;
+    @BeforeEach
+    public void setup() throws IOException {
+        when(updateCenterService.fetchUpdateCenter()).thenReturn(new UpdateCenter(
+            Map.of(),
+            Map.of(),
+            List.of()
+        ));
+    }
 
     @Test
     public void shouldBeAbleToRunSimpleProbe() throws IOException {
         final Plugin plugin = mock(Plugin.class);
         final Probe probe = mock(Probe.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
 
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probe), pluginService, updateCenterService);
         final ProbeResult expectedResult = ProbeResult.success("foo", "bar");
 
+        when(plugin.getName()).thenReturn("foo");
         when(probe.doApply(any(Plugin.class), any(ProbeContext.class))).thenReturn(expectedResult);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
         verify(probe).doApply(any(Plugin.class), any(ProbeContext.class));
@@ -82,13 +102,18 @@ class ProbeEngineTest {
         final String probeKey = "wiz";
         final Probe probe = mock(Probe.class);
         final ProbeContext ctx = mock(ProbeContext.class);
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probe), pluginService, updateCenterService);
 
+        when(plugin.getName()).thenReturn("foo");
         when(plugin.getReleaseTimestamp()).thenReturn(ZonedDateTime.now().minusDays(1));
         when(plugin.getDetails()).thenReturn(Map.of(probeKey, ProbeResult.success(probeKey, "This is good")));
         when(probe.requiresRelease()).thenReturn(true);
         when(probe.key()).thenReturn(probeKey);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
         verify(probe, never()).doApply(plugin, ctx);
@@ -96,19 +121,76 @@ class ProbeEngineTest {
     }
 
     @Test
+    public void shouldNotApplyProbeRelatedToCodeWithNoNewCode() throws IOException {
+        final Plugin plugin = mock(Plugin.class);
+        final Probe probe = mock(Probe.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+
+        when(plugin.getDetails()).thenReturn(Map.of(
+            "probe", new ProbeResult("probe", "message", ResultStatus.SUCCESS, ZonedDateTime.now().minusDays(1))
+        ));
+        when(plugin.getName()).thenReturn("foo");
+        when(probe.key()).thenReturn("probe");
+        when(probe.requiresRelease()).thenReturn(false);
+        when(probe.isSourceCodeRelated()).thenReturn(true);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
+        when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
+        probeEngine.run();
+
+        verify(probe, never()).apply(plugin, ctx);
+    }
+
+    @Test
+    public void shouldApplyProbeRelatedToCodeWithNewCommit() throws IOException {
+        final Plugin plugin = mock(Plugin.class);
+        final Probe probe = mock(Probe.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+        final ProbeResult result = new ProbeResult("probe", "message", ResultStatus.SUCCESS);
+
+        when(plugin.getDetails()).thenReturn(Map.of(
+            "probe", new ProbeResult("probe", "message", ResultStatus.SUCCESS, ZonedDateTime.now().minusDays(1))
+        ));
+        when(ctx.getLastCommitDate()).thenReturn(Optional.of(ZonedDateTime.now()));
+        when(plugin.getName()).thenReturn("foo");
+        when(probe.key()).thenReturn("probe");
+        when(probe.isSourceCodeRelated()).thenReturn(true);
+        when(probe.doApply(plugin, ctx)).thenReturn(result);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
+        when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
+        probeEngine.run();
+
+        verify(probe).apply(plugin, ctx);
+        verify(plugin).addDetails(result);
+    }
+
+    @Test
     public void shouldApplyProbeWithReleaseRequirementOnPluginWithNewReleaseAndPastResult() throws IOException {
         final String probeKey = "wiz";
         final Plugin plugin = mock(Plugin.class);
         final Probe probe = mock(Probe.class);
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probe), pluginService, updateCenterService);
+        final ProbeContext ctx = mock(ProbeContext.class);
 
+        when(plugin.getName()).thenReturn("foo");
         when(plugin.getReleaseTimestamp()).thenReturn(ZonedDateTime.now());
         when(plugin.getDetails()).thenReturn(Map.of(probeKey, new ProbeResult(probeKey, "this is ok", ResultStatus.SUCCESS, ZonedDateTime.now().minusDays(1))));
         when(probe.requiresRelease()).thenReturn(true);
         when(probe.apply(any(Plugin.class), any(ProbeContext.class)))
             .thenReturn(ProbeResult.success(probeKey, "This is also ok"));
         when(probe.key()).thenReturn(probeKey);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
         verify(probe).doApply(eq(plugin), any(ProbeContext.class));
@@ -120,16 +202,25 @@ class ProbeEngineTest {
         final String probeKey = "wiz";
         final Plugin plugin = mock(Plugin.class);
         final Probe probe = mock(Probe.class);
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probe), pluginService, updateCenterService);
+        final ProbeContext ctx = mock(ProbeContext.class);
 
-        when(plugin.getDetails()).thenReturn(Map.of(probeKey, new ProbeResult(probeKey, "this is ok", ResultStatus.SUCCESS, ZonedDateTime.now().minusDays(1))));
+        when(plugin.getName()).thenReturn("foo");
+        when(plugin.getDetails()).thenReturn(Map.of(
+            probeKey,
+            new ProbeResult(probeKey, "this is ok", ResultStatus.SUCCESS, ZonedDateTime.now().minusDays(1))
+        ));
         when(probe.requiresRelease()).thenReturn(false);
         when(probe.apply(eq(plugin), any(ProbeContext.class))).thenReturn(ProbeResult.success(probeKey, "This is also ok"));
         when(probe.key()).thenReturn(probeKey);
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
-        verify(probe).doApply(eq(plugin), any(ProbeContext.class));
+        verify(probe).doApply(plugin, ctx);
         verify(pluginService).saveOrUpdate(plugin);
     }
 
@@ -137,10 +228,16 @@ class ProbeEngineTest {
     public void shouldNotSaveErrors() throws IOException {
         final Plugin plugin = mock(Plugin.class);
         final Probe probe = mock(Probe.class);
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probe), pluginService, updateCenterService);
+        final ProbeContext ctx = mock(ProbeContext.class);
 
+        when(plugin.getName()).thenReturn("foo");
         when(probe.doApply(eq(plugin), any(ProbeContext.class))).thenReturn(ProbeResult.error("foo", "bar"));
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probe));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
         verify(plugin, never()).addDetails(any(ProbeResult.class));
@@ -150,6 +247,7 @@ class ProbeEngineTest {
     public void shouldBeAbleToGetPreviousContextResultInExecution() throws IOException {
         final Plugin plugin = spy(Plugin.class);
         final Probe probeOne = mock(Probe.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
         final Probe probeTwo = new Probe() {
             @Override
             protected ProbeResult doApply(Plugin plugin, ProbeContext ctx) {
@@ -173,12 +271,17 @@ class ProbeEngineTest {
                 return null;
             }
         };
-        final ProbeEngine probeEngine = new ProbeEngine(List.of(probeOne, probeTwo), pluginService, updateCenterService);
 
+        when(plugin.getName()).thenReturn("foo");
         when(probeOne.key()).thenReturn("foo");
         when(probeOne.doApply(any(Plugin.class), any(ProbeContext.class)))
             .thenReturn(ProbeResult.success("foo", "This is ok"));
+
+        when(probeService.getProbeContext(anyString(), any(UpdateCenter.class))).thenReturn(ctx);
+        when(probeService.getProbes()).thenReturn(List.of(probeOne, probeTwo));
         when(pluginService.streamAll()).thenReturn(Stream.of(plugin));
+
+        final ProbeEngine probeEngine = new ProbeEngine(probeService, pluginService, updateCenterService, githubConfiguration);
         probeEngine.run();
 
         verify(plugin, times(2)).addDetails(any(ProbeResult.class));
